pyproject.toml
[tool.poetry]
name = "patroacore"
version = "0.1.0"
description = "Patroa Core Backend"
authors = ["Patroa Team"]

[tool.poetry.dependencies]
python = ">=3.12"
django = "^6.0.4"
djangorestframework = "^3.17.1"
requests = "^2.33.1"
google-auth = "^2.49.2"
djangorestframework-simplejwt = "^5.5.1"
django-ratelimit = "^4.1.0"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
